[feature] 마지막 수정. float 값으로 바꾸고 매번 바뀌는 ip 값 수정하는곳 로그처리.

+++ notcompose/src/main/java/com/example/notcompose/ui/sensor/SensorActivity.kt
@@ -79,7 +79,7 @@ class SensorActivity : BaseActivity<ActivitySensorBinding>(R.layout.activity_sen
         binding.lineChart.data = lineData
     }
 
-    private fun addEntry(type: String, sensorData: Int) {
+    private fun addEntry(type: String, sensorData: Float) {
         val data = binding.lineChart.data
 
         data?.let {
@@ -89,7 +89,7 @@ class SensorActivity : BaseActivity<ActivitySensorBinding>(R.layout.activity_sen
                 data.addDataSet(set)
             }
 
-            floatTemp = sensorData.toFloat()
+            floatTemp = sensorData
             data.addEntry(Entry(set.entryCount.toFloat(), floatTemp!!), 0)
             data.notifyDataChanged()
             binding.lineChart.apply {
